+++ python/nano/test/pytorch/scripts/test-nano-pytorch.sh
@@ -7,16 +7,12 @@ export PYTORCH_NANO_TEST_DIR=${ANALYTICS_ZOO_ROOT}/python/nano/test
 wget -nv ${FTP_URI}/analytics-zoo-data/cifar-10-python.tar.gz -P ${PYTORCH_NANO_TEST_DIR}/data
 
 set -e
-echo "#Start bigdl-nano ipex tests"
-echo "#1 Start test model vision ipex"
-
-echo "#3 Start test models onnx"
+echo "# Start testing"
 start=$(date "+%s")
-python -m pytest -s ${PYTORCH_NANO_TEST_DIR}/pytorch/test_lightning.py
-now=$(date "+%s")
-time3=$((now-start))
+python -m pytest -s ${PYTORCH_NANO_TEST_DIR}/pytorch/tests/
 
-echo "Bigdl-nano ipex tests finished"
-echo "#1 model vision ipex time used:$time1 seconds"
-echo "#2 trainer ipex time used:$time2 seconds"
-echo "#3 models onnx time used:$time3 seconds"
+now=$(date "+%s")
+time=$((now-start))
+
+echo "Bigdl-nano tests finished"
+echo "Time used:$time seconds"
